Move the network location reader next to NodeObservation
_coords.py is a module of coordinate types. location_from_coords was the
one piece of logic in it, and what it returns is NodeObservation.at's type
-- a node name or a (reach, distance) pair. Both callers already import
obs.py, so the move costs no new import edge.
