+++ src/modelskill/obs.py
@@ -866,6 +866,29 @@ class VerticalObservation(Observation):
         return self._coordinate_values("z")
 
 
+def _location_from_coords(ds: xr.Dataset) -> Any:
+    """Where a network timeseries sits, as the network that produced it named it.
+
+    Returns a node name for a node, a ``(reach, distance)`` pair for a
+    breakpoint, a reach name when no distance was given, and None for data that
+    carries no network location. The value is returned as recorded, so a comparer
+    saved by an older version gives back the integer it stored.
+    """
+    if "node" in ds.coords:
+        return _scalar(ds, "node")
+    if "reach" in ds.coords:
+        reach = _scalar(ds, "reach")
+        if "distance" not in ds.coords:
+            return reach
+        return (reach, _scalar(ds, "distance"))
+    return None
+
+
+def _scalar(ds: xr.Dataset, name: str) -> Any:
+    value = np.atleast_1d(ds.coords[name].values)[0]
+    return value.item() if hasattr(value, "item") else value
+
+
 class NodeObservation(Observation):
     """Class for observations at network nodes.
